Multiple changes based on new view of driver cams
Main change - moving to use of gstreamer, rather than the install of
NodeJS and h264player.  There are some integration issues that will need
to be resolved, but the video stream that results is far superior.

Moved some sections to the end under 'Archived Sections' to remove
possible confusion.

There are still more rearrangements that are necessary.